% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPfunctions.R
\name{fitGP}
\alias{fitGP}
\title{Fit a GP model}
\usage{
fitGP(
  data = NULL,
  yd,
  xd = NULL,
  pop = NULL,
  time = NULL,
  E = NULL,
  tau = NULL,
  scaling = "global",
  initpars = NULL,
  modeprior = 1,
  rhofixed = NULL,
  predictmethod = NULL,
  datanew = NULL,
  xnew = NULL,
  popnew = NULL,
  timenew = NULL,
  ynew = NULL
)
}
\arguments{
\item{data}{A data frame, or matrix with named columns.}

\item{yd}{The response variable (required). If \code{data} is supplied, a column name
(character) of index (numeric). If \code{data} is not supplied, a numeric vector.}

\item{xd}{Predictor variables. If \code{data} is supplied, column names
(character vector) of indices (numeric vector). If \code{data} is not supplied, a numeric matrix
(or vector, if there is only one predictor variable). If \code{xd} is not supplied,
values for \code{E} and \code{tau} must be provided to construct it internally.}

\item{pop}{Identifies separate populations (optional, if not supplied, defaults to 1
population). Population values can be either numeric, character, or factor.
If \code{data} is supplied, a column name (character) of index (numeric).
If \code{data} is not supplied, a vector (numeric, character, or factor).}

\item{time}{A time index (optional, if not supplied, defaults to a numeric index).
Important: The time index is not used for model fitting (timesteps are
assumed to be evenly spaced) but supplying \code{time} will be add these values to the output table,
which may be useful for later plotting purposes. If \code{data} is supplied, a column name
(character) of index (numeric). If \code{data} is not supplied, a numeric vector.}

\item{E}{Embedding dimension. Required if \code{xd} is not supplied.}

\item{tau}{Time delay. Required if \code{xd} is not supplied.}

\item{scaling}{How the variables should be scaled (see Details). Scaling can be \code{"global"}
(default), \code{"local"} (only applicable if more than 1 pop), or \code{"none"}. Scaling is applied
to \code{yd} and each \code{xd}. For a mix of scalings, do it manually beforehand and
set scaling to \code{"none"}. All outputs are backtransformed to original scale.}

\item{initpars}{Starting values for hyperparameters (see Details) in the order
\code{c(phi,ve,tau,rho)}, in constrained (non-transformed) space. Should be a numeric
vector with length \code{ncol(xd)+3}. Defaults used if not supplied: \code{c(rep(0.1,ncol(xd)), 0.001, 1-0.001, 0.5)}}

\item{modeprior}{This value is used by the phi prior and sets the expected number of modes over the unit interval.
Defaults to 1.}

\item{rhofixed}{Fixes the rho parameter, if desired (see Details).}

\item{predictmethod}{\code{loo} (default) does leave-one-out prediction using
the training data. \code{sequential} does sequential (leave-future-out)
prediction using the training data.}

\item{datanew}{Data frame containing the same columns supplied in the
original model.}

\item{xnew}{New preditor matrix or vector. Not required if \code{datanew} is supplied.}

\item{popnew}{New population vector. Not required if \code{datanew} is supplied.}

\item{timenew}{New time vector. Not required if \code{datanew} is supplied.}

\item{ynew}{New response vector. Optional, unless \code{E} and \code{tau} were supplied in lieu of \code{xd}. Not required if \code{datanew} is supplied.}
}
\value{
A list (class GP) with the following elements:
\item{pars}{Posterior mode of hyperparameters.}
\item{parst}{Posterior mode of hyperparameters, transformed to real line (used for optimization).}
\item{grad}{Likelihood gradients of hyperparameters at posterior mode. Can be used to assess convergence.}
\item{covm}{List containing various covariance matrices and the inverse covariance matrix.}
\item{iter}{Number of iterations until convergence was reached. Currently fixed at a max of 200.}
\item{inputs}{Inputs and scaled inputs.}
\item{scaling}{Scaling information.}
\item{insampresults}{Data frame with in-sample predictions. \code{predfsd} is the standard
deviation of the GP function, \code{predsd} includes observation error.}
\item{insampfitstats}{Fit statistics for in-sample predictions.}
\item{outsampresults}{Data frame with out-of-sample predictions (if requested). \code{predfsd} is the standard
deviation of the GP function, \code{predsd} includes observation error.}
\item{outsampfitstats}{Fit statistics for out-of-sample predictions (if requested).
Only computed if using \code{"loo"} or \code{"sequential"}, if \code{yd} is found in \code{datanew},
or if \code{ynew} supplied (i.e. if the observed values are known).}
}
\description{
Fits a GP model with separable length scales and automatic relevance determination (ARD).
Also fits a hierarchical version of the GP model if distinct populations are indicated
using \code{pop}.
There are several ways to specify the training data:
\enumerate{
\item data frame \code{data}, plus column names or indices for \code{yd} and \code{xd}
\item data frame \code{data}, plus column name or index for \code{yd}, and values for \code{E} and \code{tau}
\item vector \code{yd}, plus matrix or vector \code{xd}
\item vector \code{yd}, plus values for E and tau
}
Arguments \code{pop} and \code{time} are optional in all of the above cases.
They should be column names or indices in cases 1 and 2, vectors in cases 3 and 4.
If \code{pop} and \code{time} are supplied, they should NOT contain missing values.
This function will also, optionally, produce predictions.
See Details for more information.
}
\details{
The data are assumed to be should be in time order. This is particularly important if the E/tau option
or \code{predictmethod="sequential"} is used.

\strong{Missing values:}

Missing values for \code{yd} and \code{xd} are allowed. Any rows containing
missing values will be excluded from the model fitting procedure (reinserted
as NAs in the output). If \code{pop} and \code{time} are supplied, they
should NOT contain missing values.

\strong{Hyperparameters:}

The model uses the following covariance function:

(Insert equation here.)

There is one inverse length scale \code{phi} estimated for each predictor variable.
Values near 0 indicate that the predictor has little influence on the response variable,
and it was likely dropped by ARD. Higher values of \code{phi} indicate greater nonlinearity.

The estimated variance terms are \code{ve} (observation variance) and
\code{tau} (pointwise function variance).

Hyperparameter \code{rho} is the dynamic
correlation in the hierarchical GP model, indicating similarity of dynamics
across populations. If there is only 1 population (e.g. if \code{pop} is not
supplied), \code{rho} is irrelevant (not used by the model) and will revert
to the mode of the prior (~0.5). The value of \code{rho} can be fixed to any
value from 0.0001 to 1 using \code{rhofixed}, otherwise it is estimated from the data.
See Munch et al. (2017) for more details.

\strong{Scaling:}

The model priors assume the response variable and all predictor
variables have approximately zero mean and unit variance, therefore it is
important to scale the data properly when fitting these models. For
convenience, this function will scale the input data automatically by
default, and backtransform output to the original scale. Automatic scaling
can be \code{"global"} (default), or \code{"local"}. The latter will scale
variables within (as opposed to across) populations, so is obviously only
applicable if there is more than 1 population. You can also scale the data
yourself beforehand in whatever manner you wish and set
\code{scaling="none"}. In this case, you will obviously have to do the
backtransformation yourself.

\strong{Predictions:}

There are several options for producing predictions:
\enumerate{
\item Use \code{predictmethod="loo"} for leave-one-out prediction using the training data
\item Use \code{predictmethod="sequential"} for sequential (leave-future-out) prediction using the training data
\item If data frame \code{data} was supplied, supply data frame \code{datanew} containing same column names.
Column for \code{yd} is optional, unless \code{E} and \code{tau} were supplied in lieu of \code{xd}.
\item If vectors/matrices were supplied for \code{yd}, \code{xd}, etc, equivalent vector/matrices \code{xnew},
\code{popnew} (if \code{pop} was supplied), and \code{timenew} (optional).
\code{ynew} is optional, unless \code{E} and \code{tau} were supplied in lieu of \code{xd}.
}

After fitting, predictions can also be made using \code{\link{predict.GP}}.
Predictions can plotted using \code{\link{plot.GP}}.
Get conditional responses using \code{\link{getconditionals}}.
}
\examples{
yrand=rnorm(20)
testgp=fitGP(yd=yrand,E=2,tau=1,predictmethod = "loo")
}
\references{
Munch, S. B., Poynor, V., and Arriaza, J. L. 2017. Circumventing structural uncertainty:
a Bayesian perspective on nonlinear forecasting for ecology. Ecological Complexity, 32: 134.
}
\seealso{
\code{\link{predict.GP}}, \code{\link{plot.GP}}, \code{\link{getconditionals}}
}
