% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPfisheries.R
\name{fitGP_fish}
\alias{fitGP_fish}
\title{Fit a "fisheries" GP model}
\usage{
fitGP_fish(
  data,
  y,
  m,
  h,
  z = NULL,
  pop = NULL,
  time = NULL,
  scaling = c("global", "local", "none"),
  initpars = NULL,
  modeprior = 1,
  rhofixed = NULL,
  rhomatrix = NULL,
  predictmethod = NULL,
  newdata = NULL,
  xname = "escapement"
)
}
\arguments{
\item{data}{A data frame, required.}

\item{y}{The response variable (required).}

\item{m}{Lags of the response variable (required).}

\item{h}{Lags of the variable to be multipled by b and subtracted from m (required).}

\item{z}{Other predictor variables, e.g. covariates, that are unmodified (optional).}

\item{pop}{Identifies separate populations (optional, if not supplied, defaults to 1
population). Population values can be either numeric, character, or factor.}

\item{time}{The time index (recommended). If not supplied, defaults to a numeric index.}

\item{scaling}{How the variables should be scaled (see Details). Scaling can be \code{"global"}
(default), \code{"local"} (only applicable if more than 1 pop), or \code{"none"}. Scaling is applied
to \code{y} and each \code{x}. For a mix of scalings, do it manually beforehand and
set scaling to \code{"none"}. All outputs are backtransformed to original scale.}

\item{initpars}{Starting values for hyperparameters (see Details) in the order
\code{c(phi,ve,sigma2,rho)}, in constrained (non-transformed) space. Should be a numeric
vector with length \code{ncol(x)+3}. Defaults used if not supplied: \code{c(rep(0.1,ncol(x)), 0.001, 1-0.001, 0.5)}}

\item{modeprior}{This value is used by the phi prior and sets the expected number of modes over the unit interval.
Defaults to 1.}

\item{rhofixed}{Fixes the rho parameter, if desired (see Details).}

\item{rhomatrix}{Symmetrical square matrix of fixed pairwise rho values to use, with 1's on the diagonal.
The rows and columns should be named with the population identifier.
The output of \code{\link{getrhomatrix}} can be used here.
All populations in the dataset must appear in \code{rhomatrix}. Populations in
\code{rhomatrix} that are not in the dataset are allowed and will not be used.}

\item{predictmethod}{Using the training data, \code{loo} does leave-one-out prediction, \code{lto} does
leave-timepoint-out prediction, and \code{sequential} does sequential (leave-future-out)
prediction.}

\item{newdata}{Data frame containing the same columns supplied in the
original model.}

\item{xname}{What the composite variable m-bh should be called. Defaults to "escapement".}
}
\value{
A list (class GP and GPpred) with the same elements as fitGP with additonal
element \code{b}, data_x, and names of m, h, z stored under inputs.
}
\description{
Fits an alternate version of the GP model with one additonal parameter (catchability),
designed for use in fisheries applications.
}
\details{
This fits a GP model of the form
y=f(m-bh,z)

Where y is CPUE, m are lags of CPUE, h are lags of harvest, b is catchability (scalar),
and z are optional covariates. The dimensions of m and h must match.

Parameter b is found using \code{optimize} applied to the posterior likelihood.

Using this method requires the use of \code{data} with pre-generated lags
(option A1 in \code{\link{fitGP}}).
}
\keyword{functions}
