% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotsConditionals.R
\name{getconditionals}
\alias{getconditionals}
\title{Get conditional responses from a GP model}
\usage{
getconditionals(fit, plot = T)
}
\arguments{
\item{fit}{Output from \code{fitGP}.}

\item{plot}{Produce a plot, or not (logical).}
}
\value{
A data frame.
}
\description{
This will obtain responses to each predictor variable across its range of
observed values, conditional on all other predictor variables being set to 0
(their mean value, if data are scaled properly). If automatic scaling is
used, predictors and responses are backtransformed to their original scale.
If the model includes multiple populations, responses are obtained for each
population.
}
