% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPfunctions.R
\name{getR2}
\alias{getR2}
\title{Calculate R-squared}
\usage{
getR2(obs, pred)
}
\arguments{
\item{obs}{Vector of observed values.}

\item{pred}{Vector of predicted values.}
}
\value{
The R-squared value.
}
\description{
Calculates an R-squared value. Vectors for observed and predicted values
must be the same length. Missing values are allowed. Any observed-predicted
pairs with missing values are removed before computations are done.
}
\details{
Returned R-squared might be negative. This indicates that the prediction is
worse than using the mean.
}
