% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPfunctions.R
\name{getR2pop}
\alias{getR2pop}
\title{Calculate R-squared by population (and other stuff)}
\usage{
getR2pop(obs, pred, pop, type = c("within", "centered", "scaled"))
}
\arguments{
\item{obs}{Vector of observed values.}

\item{pred}{Vector of predicted values.}

\item{pop}{Vector of pop identifiers.}

\item{type}{The type of R-squared to calculate: \code{"within"} (default) gets
within-population, \code{"centered"} gets across-population but with local means removed,
\code{"scaled"} get across-population but with local scaling (local means removed and
rescaled to unit variance).}
}
\value{
If \code{type="within"}, a list of 2 named vectors with R-squared and rmse for
each populaiton. Otherwise a scalar (the R-squared).
}
\description{
Calculates R-squared values for each population. Can also compute modified
total R-squared values, with data centered and/or scaled within populations.
This can provide a more useful total fit estimate if the populations have
very different local means.
}
\details{
Returned R-squared might be negative. This indicates that the prediction is
worse than using the mean.
}
\seealso{
\code{\link{getR2}}
}
\keyword{functions}
