% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhomatrix.R
\name{getrhomatrix}
\alias{getrhomatrix}
\title{Get pairwise dynamic correlation matrix}
\usage{
getrhomatrix(
  data = NULL,
  y,
  x = NULL,
  pop,
  time = NULL,
  E = NULL,
  tau = NULL,
  scaling = c("global", "local", "none"),
  initpars = NULL,
  modeprior = 1,
  augdata = NULL
)
}
\arguments{
\item{data}{A data frame, or matrix with named columns.}

\item{y}{The response variable (required). If \code{data} is supplied, a column name
(character) of index (numeric). If \code{data} is not supplied, a numeric vector.}

\item{x}{Predictor variables. If \code{data} is supplied, column names
(character vector) of indices (numeric vector). If \code{data} is not supplied, a numeric matrix
(or vector, if there is only one predictor variable). If \code{x} is not supplied,
values for \code{E} and \code{tau} must be provided to construct it internally.}

\item{pop}{Identifies separate populations (optional, if not supplied, defaults to 1
population). Population values can be either numeric, character, or factor.
If \code{data} is supplied, a column name (character) of index (numeric).
If \code{data} is not supplied, a vector (numeric, character, or factor).}

\item{time}{A time index (optional, if not supplied, defaults to a numeric index).
Important: The time index is not used for model fitting (timesteps are
assumed to be evenly spaced) but supplying \code{time} will be add these values to the output table,
which may be useful for later plotting purposes. If \code{data} is supplied, a column name
(character) of index (numeric). If \code{data} is not supplied, a numeric vector.}

\item{E}{Embedding dimension. If supplied, will be used to constuct lags of \code{x} (or
lags of \code{y} if \code{x} is not supplied).}

\item{tau}{Time delay. If supplied, will be used to constuct lags of \code{x} (or
lags of \code{y} if \code{x} is not supplied).}

\item{scaling}{How the variables should be scaled (see Details). Scaling can be \code{"global"}
(default), \code{"local"} (only applicable if more than 1 pop), or \code{"none"}. Scaling is applied
to \code{y} and each \code{x}. For a mix of scalings, do it manually beforehand and
set scaling to \code{"none"}. All outputs are backtransformed to original scale.}

\item{initpars}{Starting values for hyperparameters (see Details) in the order
\code{c(phi,ve,sigma2,rho)}, in constrained (non-transformed) space. Should be a numeric
vector with length \code{ncol(x)+3}. Defaults used if not supplied: \code{c(rep(0.1,ncol(x)), 0.001, 1-0.001, 0.5)}}

\item{modeprior}{This value is used by the phi prior and sets the expected number of modes over the unit interval.
Defaults to 1.}

\item{augdata}{A data frame with augmentation data (see Details).}
}
\value{
A square matrix of pairwise dynamic correlations between all populations. Rows
and columns are named with population names.
}
\description{
Fits a GP model to each pair of populations and constructs a matrix of pairwise
dynamic correlation (\code{rho}) values. The matrix can be passed to \code{fitGP}
under argument \code{rhomatrix}.
}
\details{
The \code{pop} argument is \emph{required} and there must be more than 1 population.
Function scales data prior to subsetting to each pair of populations.
}
\references{
Rogers, T. L. and Munch, S. B. 2020. Hidden similarities in the dynamics
of weakly synchronous marine metapopulation. Proceedings of the National Academy of
Sciences 117(1):479-485
}
\seealso{
\code{\link{fitGP}}
}
\keyword{functions}
