% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsubset.R
\name{getsubset}
\alias{getsubset}
\title{Subset training data using kmeans clustering}
\usage{
getsubset(
  xds,
  yds,
  nclust,
  clustsize = NULL,
  subtotal = NULL,
  returnclusters = FALSE
)
}
\arguments{
\item{xds}{Predictor variables as a matrix. These should be scaled! Can contain
missing values.}

\item{yds}{Response variable as a vector. This is only used to exclude rows
with missing values for y; it is not used for clustering.}

\item{nclust}{The number of clusters.}

\item{clustsize}{The number of observations per cluster.}

\item{subtotal}{The total number of observations in the subset. The
number of observations per cluster will be subtotal/nclust. If subtotal
is not an exact multiple of nclust, then the remainder is assigned to
random clusters.}

\item{returnclusters}{Return the cluster assignments for each data point.
Potentially useful for diagnostics.}
}
\value{
A vector of indices indicating the rows of xds and yds that should be included,
which can be used for subsetting. If returnclusters=TRUE, a list with a vector of indices
and a vector of cluster assignments.
}
\description{
Reduces the size of the training dataset for 'big data' GP applications. This
function is used internally in \code{fitGP} when \code{kmsubset=TRUE} (after
data scaling but before subsetting out missing data - rows not in the subset
are treated as 'missing'). However, this function can also be used on its own for
diagnostic purposes, or to pre-subset the training data if desired.
}
\details{
For a given training dataset, uses kmeans to divide the predictor rows into
clusters, and then subsamples (without replacement) from each of those clusters.
The goal is to select points that are reasonably distributed across the attractor,
which appears to work better than points selected completely at random (nclust=1).

The number of clusters must be specified, along with either the number of points per
cluster or the total size of the subsample.

If you want to use this function, note that xds (x) must be a matrix and the predictors
must be scaled. There can be missing values: the clustering and subsampling is only done
for complete rows (no missing x or y values). In other words, the indices that come out
will remove both incomplete rows and the rows not in the subset.
}
\keyword{functions}
