% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HastPow3sp}
\alias{HastPow3sp}
\title{Simulated time series from 3 species Hastings-Powell model}
\format{
A data frame with 501 rows and 4 variables:
\describe{
\item{Time}{Time index}
\item{X}{Abundance of species X}
\item{Y}{Abundance of species Y}
\item{Z}{Abundance of species Z}
}
}
\source{
The simulation code is in data-raw on the GPEDM Github page.
}
\usage{
HastPow3sp
}
\description{
Contains time series simulated from the continuous time 3 species Hastings-Powell model:
\deqn{\frac{dX}{dt}=X(1-X)-aXY/(1+bX)}
\deqn{\frac{dY}{dt}=aXY/(1+bx)-cYZ/(1+dY)-mY}
\deqn{\frac{dZ}{dt}=cYZ/(1+dY)-\mu Z}
Parameters are \eqn{a=5, b=3, c=0.1, d=2, m=0.4, \mu=0.01}.
Initial conditions are \eqn{X(0)=0.5, Y(0)=0.1, Z(0)=9}.
Integrated for 500 time steps. Both have \eqn{r=1.2} and \eqn{s=0.05} and time series length of 50,
with random starting values and the first 200 iterates discarded.
}
\keyword{datasets}
