% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPfunctions.R
\name{makelags}
\alias{makelags}
\title{Generate delay vectors}
\usage{
makelags(yd, pop = NULL, E, tau, yname = NULL)
}
\arguments{
\item{yd}{A vector containing a time series. Alternatively, a matrix or data
frame where each column is a time series. In the case of multiple time
series, lags will be generated for each variable.}

\item{pop}{A vector of populations (optional).}

\item{E}{Embedding dimension.}

\item{tau}{Time delay.}

\item{yname}{Optional, name of the variable if yd is a vector, or variables
if yd is a matrix or dataframe with unnamed columns. Otherwise not used.
Defaults to "var" if NULL.}
}
\value{
A matrix with named columns, the appended number indicating tau. If
\code{yd} has named columns, named columns in the lag matrix will match.
}
\description{
Create a lag matrix for a time series (or suite of time series) using given E
and tau values.
}
\examples{
yrand <- rnorm(20)
site <- rep(c("a","b"),each=10)
dfrand <- data.frame(firstvar=rnorm(20),secondvar=rnorm(20))
lags1 <- makelags(yrand,E=3,tau=1)
lags2 <- makelags(dfrand,E=2,tau=2)
lags3 <- makelags(dfrand,pop=site,E=2,tau=1)

}
