% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_iter.R
\name{msy_wrapper}
\alias{msy_wrapper}
\title{Wrapper for computing MSY from a "fisheries model"}
\usage{
msy_wrapper(model, newdata, hratevec, tsave)
}
\arguments{
\item{model}{A \code{fitGP_fish} model.}

\item{newdata}{Data frame for iterated prediction, such as made with
\code{makelags_iter}.}

\item{hratevec}{Vector of harvest rates}

\item{tsave}{Number of timepoints to average over. Should not exceed nrow(newdata)}
}
\description{
Wrapper for computing MSY from a "fisheries model"
}
\keyword{functions}
