% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotsConditionals.R
\name{plot.GPpred}
\alias{plot.GPpred}
\title{Plot predictions from a GP model}
\usage{
\method{plot}{GPpred}(x, plotinsamp = F, ...)
}
\arguments{
\item{x}{Output from \code{fitGP} or \code{predict.GP}.}

\item{plotinsamp}{Plot the in-sample results. Defaults to out-of-sample
results if available.}

\item{...}{Other args (not used).}
}
\description{
Plots observed and predicted values from a GP model. If the model includes
multiple populations, separate plots are produced for each population. The
standard deviations plotted are predfsd.
}
\keyword{functions}
