% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotsConditionals.R
\name{plot.NStest}
\alias{plot.NStest}
\title{Plot results from an S-map nonstationarity test}
\usage{
\method{plot}{NStest}(x, ...)
}
\arguments{
\item{x}{Output from \code{Smap_NStest}.}

\item{...}{Other graphical parameters for the time series plot, like 'main'.}
}
\description{
Makes some plots of results from the nonstationarity test.
}
\keyword{functions}
