% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotsConditionals.R
\name{plot.Smappred}
\alias{plot.Smappred}
\title{Plot predictions from an S-map model}
\usage{
\method{plot}{Smappred}(x, plotloo = F, ...)
}
\arguments{
\item{x}{Output from \code{fitSmap} or \code{predict.Smap}.}

\item{plotloo}{Plot the leave-one-out results. Defaults to other out-of-sample
results if available.}

\item{...}{Other args (not used).}
}
\description{
Plots observed and predicted values from an S-map model. If the model includes
multiple populations, separate plots are produced for each population. Plots
leave-one-out predictions, unless other predictions are available.
}
\keyword{functions}
