% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhomatrix.R
\name{posdef}
\alias{posdef}
\title{Find nearest positive definite matrix}
\usage{
posdef(mat)
}
\arguments{
\item{mat}{A symmetrical square matrix with 1's on diagonal.}
}
\value{
An approximate version of \code{mat} that is positive definite. Row and column names are retained.
}
\description{
If the output matrix from \code{\link{getrhomatrix}} is not positive
definite, this will find an approximate matrix that is using \code{Matrix::nearPD()}
with \code{corr=TRUE}.
}
\seealso{
\code{\link{getrhomatrix}}
}
\keyword{functions}
