% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPfunctions.R
\name{predict.GP}
\alias{predict.GP}
\title{Get predictions from a GP model}
\usage{
\method{predict}{GP}(
  object,
  predictmethod = "loo",
  datanew = NULL,
  xnew = NULL,
  popnew = NULL,
  timenew = NULL,
  ynew = NULL
)
}
\arguments{
\item{object}{Output from \code{fitGP}.}

\item{predictmethod}{\code{loo} (default) does leave-one-out prediction using
the training data. \code{sequential} does sequential (leave-future-out)
prediction using the training data.}

\item{datanew}{Data frame containing the same columns supplied in the
original model.}

\item{xnew}{New preditor matrix or vector. Not required if \code{datanew} is supplied.}

\item{popnew}{New population vector. Not required if \code{datanew} is supplied.}

\item{timenew}{New time vector. Not required if \code{datanew} is supplied.}

\item{ynew}{New response vector. Optional, unless \code{E} and \code{tau} were supplied in lieu of \code{xd}. Not required if \code{datanew} is supplied.}
}
\value{
A list with the following elements:
\item{outsampresults}{Data frame with out-of-sample predictions (if requested). \code{predfsd} is the standard
deviation of the GP function, \code{predsd} includes observation error.}
\item{outsampfitstats}{Fit statistics for out-of-sample predictions.
Only computed if using \code{"loo"} or \code{"sequential"}, if \code{yd} is found in \code{datanew},
or if \code{ynew} supplied (i.e. if the observed values are known).}
}
\description{
Obtain predictions from a fitted GP model. There are several options:
\enumerate{
\item Use \code{predictmethod="loo"} for leave-one-out prediction using the training data
\item Use \code{predictmethod="sequential"} for sequential (leave-future-out) prediction using the training data
\item If data frame \code{data} was supplied, supply data frame \code{datanew} containing same column names.
Column for \code{yd} is optional, unless \code{E} and \code{tau} were supplied in lieu of \code{xd}.
\item If vectors/matrices were supplied for \code{yd}, \code{xd}, etc, equivalent vector/matrices \code{xnew},
\code{popnew} (if \code{pop} was supplied), and \code{timenew} (optional).
\code{ynew} is optional, unless \code{E} and \code{tau} were supplied in lieu of \code{xd}.
}
}
