% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smapfunctions.R
\name{predict.Smap}
\alias{predict.Smap}
\title{Get predictions from an S-map model}
\usage{
\method{predict}{Smap}(
  object,
  predictmethod = NULL,
  newdata = NULL,
  xnew = NULL,
  popnew = NULL,
  timenew = NULL,
  ynew = NULL,
  ...
)
}
\arguments{
\item{object}{Output from \code{fitGP}.}

\item{predictmethod}{Using the training data, \code{lto} does
leave-timepoint-out prediction, and \code{sequential} does sequential (leave-future-out)
prediction.}

\item{newdata}{Data frame containing the same columns supplied in the
original model.}

\item{xnew}{New predictor matrix or vector. Not required if \code{newdata} is supplied.}

\item{popnew}{New population vector. Not required if \code{newdata} is supplied.}

\item{timenew}{New time vector. Not required if \code{newdata} is supplied.}

\item{ynew}{New response vector. Optional, unless \code{E} and \code{tau} were supplied in
lieu of \code{x}. Not required if \code{newdata} is supplied.}

\item{...}{Other args (not used).}
}
\value{
A list (class Smappred) with the following elements:
\item{outsampresults}{Data frame with out-of-sample predictions.}
\item{outsampcoefs}{Data frame with S-map coefficients for the out-of-sample predictions. The last column is the intercept.}
\item{outsampfitstats}{Fit statistics for out-of-sample predictions.
Only computed if using \code{"lto"} or \code{"sequential"}, if \code{y} is found in \code{newdata},
or if \code{ynew} supplied (i.e. if the observed values are known).}
\item{outsampfitstatspop}{If >1 population, fit statistics for out-of-sample predictions by population.}
}
\description{
Obtain predictions from a fitted S-map model.
}
\details{
The s-map routine automatically does leave-one-out prediction, which is included in the
output, so there is no separate \code{predictmethod="loo"}.

There are several additional options for producing predictions:
\enumerate{
\item Use \code{predictmethod="lto"} for leave-timepoint-out prediction using the training data. This will leave
out values with the same time index across multiple populations, rather than each individual datapoint.
If there is only one population, \code{"lto"} will be equivalent to \code{"loo"}.
\item Use \code{predictmethod="sequential"} for sequential (leave-future-out) prediction using the training data.
\item If data frame \code{data} was supplied, supply data frame \code{newdata} containing same column names.
Column for \code{y} is optional, unless \code{E} and \code{tau} were supplied in lieu of \code{x}.
\item If vectors/matrices were supplied for \code{y}, \code{x}, etc, equivalent vector/matrices \code{xnew},
\code{popnew} (if \code{pop} was supplied), and \code{timenew} (optional).
\code{ynew} is optional, unless \code{E} and \code{tau} were supplied in lieu of \code{x}.
}
}
\keyword{functions}
