% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_seq.R
\name{predict_seq}
\alias{predict_seq}
\title{Get sequentially updated predictions from a GP model}
\usage{
predict_seq(object, newdata, restart.initpars = F)
}
\arguments{
\item{object}{Output from \code{fitGP}.}

\item{newdata}{Data frame containing the same columns supplied in the
original model.}

\item{restart.initpars}{Set initpars to pars of the previous model (F, default), or restart
with the default initspars each time (T). Starting at the last values can reduce the number
of iterations required, saving time, but might trap you in a local minimum.}
}
\value{
A list (class GP and GPpred) with the same elements as \code{\link{fitGP}}.
The model information will be for the final model including all of the original training
data plus \code{newdata}. The \code{outsampresults} and \code{outsampfitstats} will be
for the sequentially updated predictions.
}
\description{
Obtains predictions for \code{newdata} given a GP model, but sequentially
adds each new observation to the training data and refits the model. This
simulates a real-time forecasting application.
}
\details{
In comparison to \code{\link{predict.GP}}, this prediction method is similar
to \code{predictmethod="sequential"} (leave-future-out), but in this case,
the future time points are not included in the training data. It is also
similar to the train/test split using \code{newdata}, but in this case, the
training data and model are updated with each timestep.

Using this method requires the use of \code{data} with pre-generated lags
(option A1 in \code{\link{fitGP}}), that a \code{time} column is specified,
that \code{newdata} has exactly the same columns as \code{data}, and that
\code{newdata} contains observed values.

The original \code{data} data frame is obtained from the global environment,
so results will be not be accurate if it is changed. Also, if for whatever
reason you want to use \code{update} on the output of \code{predict_seq}, it
will probably not work unless you respecify \code{data} and
\code{initpars}.
}
\keyword{functions}
