% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shrimp}
\alias{shrimp}
\title{Brown shrimp in the Gulf of Mexico}
\format{
A data frame with 297 rows and 3 variables:
\describe{
\item{year}{Time index}
\item{zone}{Statistical zone (region)}
\item{cpue}{Catch per unit effort}
}
}
\source{
SEAMAP Summer Groundfish Trawl Survey
}
\usage{
shrimp
}
\description{
Annual catch-per-unit effort data for brown shrimp in the Gulf of Mexico from
the Southeast Area Monitoring and Assessment Program (SEAMAP) Summer
Groundfish Trawl Survey for 9 different statistical zones from 1987 to 2019
(33 years).
}
\references{
Tsai CH, Munch SB, Masi MD, and Pollack AG. 2022. Predicting
nonlinear dynamics of short-lived penaeid shrimp species in the Gulf of
Mexico. Canadian Journal of Fisheries and Aquatic Sciences.
https://doi.org/10.1139/cjfas-2022-0029
}
\keyword{datasets}
