% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smapfunctions.R
\name{Smap_NStest}
\alias{Smap_NStest}
\title{S-map test for nonstationarity}
\usage{
Smap_NStest(
  data,
  y,
  covar = NULL,
  pop = NULL,
  time,
  Emin = 1,
  Emax,
  tau = 1,
  thetafixed = NULL,
  exclradius = 0,
  nsvar = time,
  scaling = c("global", "local", "none"),
  initpars = NULL,
  augdata = NULL,
  summaryonly = TRUE
)
}
\arguments{
\item{data}{A data frame, or matrix with named columns.}

\item{y}{The response variable (required). If \code{data} is supplied, a column name
(character) of index (numeric). If \code{data} is not supplied, a numeric vector.}

\item{covar}{Any additional predictors (these will not be lagged)}

\item{pop}{Identifies separate populations (optional, if not supplied, defaults to 1
population). Population values can be either numeric, character, or factor.
If \code{data} is supplied, a column name (character) of index (numeric).
If \code{data} is not supplied, a vector (numeric, character, or factor).}

\item{time}{A time index (optional, if not supplied, defaults to a numeric index).
Important: The time index is not used for model fitting (timesteps are
assumed to be evenly spaced) but supplying \code{time} will be add these values to the output table,
which may be useful for later plotting purposes. If \code{data} is supplied, a column name
(character) of index (numeric). If \code{data} is not supplied, a numeric vector.}

\item{Emin}{Minimum E, defaults to 1.}

\item{Emax}{Maximum E (required). Should generally not exceed sqrt(n).}

\item{tau}{Time delay. Defaults to 1.}

\item{thetafixed}{Fixes the theta parameter, if desired (see Details).}

\item{exclradius}{Exclusion radius. How many adjacent time steps on either side of the focal
point should be excluded when fitting the local linear model? Defaults to 0 (none).
This setting is retained when making other types of predictions. Must be a positive integer.}

\item{nsvar}{If \code{ns=TRUE}, the variable supplied under \code{time} is used by default.
If for some reason you want to use a different variable, supply it here (column name of
\code{data}).}

\item{scaling}{How the variables should be scaled (see Details). Scaling can be \code{"global"},
\code{"local"} (only applicable if more than 1 pop), or \code{"none"} (default). Scaling is applied
to \code{y} and each \code{x}. All outputs are backtransformed to original scale.}

\item{initpars}{Starting values for hyperparameters (see Details) in the order
\code{c(theta,delta)}. Should be a numeric vector with length 2. Defaults used
if not supplied: \code{c(0,0)}.}

\item{augdata}{A data frame with augmentation data (see Details).}

\item{summaryonly}{Return just the summary tables, not the lists of model output for
each value of E.}
}
\value{
A list (class NStest) with the following elements:
\item{delta_agg}{The nonstationarity test statistic.}
\item{bestR2}{Best loo R2 value of all the models fit.}
\item{results}{Data frame of hyperparameters, logL, df, R2, and weights for stationary and
nonstationary model at each value of E.}
\item{series}{Data frame of time and y (for plotting purposes).}
\item{mods}{If \code{summaryonly=FALSE}, list of all stationary models.}
\item{mods_NS}{If \code{summaryonly=FALSE}, list of all nonstationary models.}
}
\description{
loop over values of E, fitSmap model, return table and delta agg value
}
