% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotsConditionals.R
\name{summary.GP}
\alias{summary.GP}
\title{Print a summary of a GP model}
\usage{
\method{summary}{GP}(object)
}
\arguments{
\item{object}{Output from \code{fitGP}.}
}
\description{
Prints out some key information including hyperparameters and fit stats.
}
\keyword{functions}
