% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotsConditionals.R
\name{summary.Smap}
\alias{summary.Smap}
\title{Print a summary of an S-map model}
\usage{
\method{summary}{Smap}(object, ...)
}
\arguments{
\item{object}{Output from \code{fitSmap}.}

\item{...}{Other args (not used).}
}
\description{
Prints out some key information including hyperparameters and fit stats.
}
\keyword{functions}
