% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{thetalog2pop}
\alias{thetalog2pop}
\title{Simulated time series from 2 theta-logistic populations}
\format{
A data frame with 100 rows and 3 variables:
\describe{
\item{Time}{Time index.}
\item{Population}{Population ID.}
\item{Abundance}{Abundance value.}
}
}
\source{
The simulation code is in data-raw on the GPEDM Github page.
}
\usage{
thetalog2pop
}
\description{
Contains time series simulated from the discrete time theta-logisic model with noise:
\deqn{x_{t+1}=x_{t}e^{r(1-x_t^\theta)}e^{s\epsilon}}
\deqn{\epsilon \sim N(0,1)}
PopA has \eqn{\theta=3.5} and PopB has \eqn{\theta=2.5}.
Both have \eqn{r=1.2} and \eqn{s=0.05} and time series length of 50,
with random starting values and the first 200 iterates discarded.
}
\keyword{datasets}
