% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trawl}
\alias{trawl}
\title{NEFSC Fall Bottom Trawl Data (subset)}
\format{
A data frame with 176 rows and 5 variables:
\describe{
\item{YEAR}{Time index}
\item{REGION}{Region: GME = Gulf of Maine, GEO = George's Bank,
SNE = Southern New England, MAB = Mid-Atlantic Bight. This is
formatted as a factor (north to south).}
\item{shortfin_squid}{CPUE of northern shortfin squid (\emph{Illex illecebrosus})}
\item{longfin_squid}{CPUE of longfin squid (\emph{Loligo pealeii})}
\item{silver_hake}{CPUE of silver hake (\emph{Merluccius bilinearis})}
}
}
\source{
https://www.fisheries.noaa.gov/inport/item/22560
}
\usage{
trawl
}
\description{
Catch-per-unit effort data for 3 species from the NOAA Northeast
Fisheries Science Center Fall Bottom Trawl Survey for 4 regions
from 1973 to 2016 (44 years).
}
\keyword{datasets}
