% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPfunctions.R
\name{fitGP}
\alias{fitGP}
\title{Fit a GP model}
\usage{
fitGP(
  data = NULL,
  y,
  x = NULL,
  pop = NULL,
  time = NULL,
  E = NULL,
  tau = NULL,
  scaling = c("global", "local", "none"),
  initpars = NULL,
  modeprior = 1,
  fixedpars = NULL,
  rhofixed = NULL,
  rhomatrix = NULL,
  augdata = NULL,
  predictmethod = NULL,
  newdata = NULL,
  xnew = NULL,
  popnew = NULL,
  timenew = NULL,
  ynew = NULL,
  returnGPgrad = FALSE,
  exclradius = 0
)
}
\arguments{
\item{data}{A data frame, or matrix with named columns.}

\item{y}{The response variable (required). If \code{data} is supplied, a column name
(character) or index (numeric). If \code{data} is not supplied, a numeric vector.}

\item{x}{Predictor variables. If \code{data} is supplied, column names
(character vector) or indices (numeric vector). If \code{data} is not supplied, a numeric matrix
(or vector, if there is only one predictor variable). If \code{x} is not supplied,
values for \code{E} and \code{tau} must be provided to construct it internally.}

\item{pop}{Identifies separate populations (optional, if not supplied, defaults to 1
population). Population values can be either numeric, character, or factor.
If \code{data} is supplied, a column name (character) or index (numeric).
If \code{data} is not supplied, a vector (numeric, character, or factor).}

\item{time}{A time index (optional, if not supplied, defaults to a numeric index).
Important: The time index is not used for model fitting (timesteps are
assumed to be evenly spaced) but supplying \code{time} will be add these values to the output table,
which may be useful for later plotting purposes. If \code{data} is supplied, a column name
(character) or index (numeric). If \code{data} is not supplied, a numeric vector.}

\item{E}{Embedding dimension. If supplied, will be used to constuct lags of \code{x} (or
lags of \code{y} if \code{x} is not supplied).}

\item{tau}{Time delay. If supplied, will be used to constuct lags of \code{x} (or
lags of \code{y} if \code{x} is not supplied).}

\item{scaling}{How the variables should be scaled (see Details). Scaling can be \code{"global"}
(default), \code{"local"} (only applicable if more than 1 pop), or \code{"none"}. Scaling is applied
to \code{y} and each \code{x}. For a mix of scalings, do it manually beforehand and
set scaling to \code{"none"}. All outputs are backtransformed to original scale.}

\item{initpars}{Starting values for hyperparameters (see Details) in the order
\code{c(phi,ve,sigma2,rho)}, in constrained (non-transformed) space. Should be a numeric
vector with length \code{ncol(x)+3}. Defaults used if not supplied: \code{c(rep(0.1,ncol(x)), 0.001, 1-0.001, 0.5)}}

\item{modeprior}{This value is used by the phi prior and sets the expected number of modes over the unit interval.
Defaults to 1.}

\item{fixedpars}{Fixes values of the hyperparameters phi, ve, and sigma2 (if desired). Should be a numeric
vector with length \code{ncol(x)+2} in the order \code{c(phi,ve,sigma2)} in constrained
(non-transformed) space. Hyperparameters to estimate should be input as NA. The value of rho is
fixed separately using argument \code{rhofixed}.}

\item{rhofixed}{Fixes the rho parameter, if desired (see Details).}

\item{rhomatrix}{Symmetrical square matrix of fixed pairwise rho values to use, with 1's on the diagonal.
The rows and columns should be named with the population identifier.
The output of \code{\link{getrhomatrix}} can be used here.
All populations in the dataset must appear in \code{rhomatrix}. Populations in
\code{rhomatrix} that are not in the dataset are allowed and will not be used.}

\item{augdata}{A data frame with augmentation data (see Details).}

\item{predictmethod}{Using the training data, \code{loo} does leave-one-out prediction, \code{lto} does
leave-timepoint-out prediction, and \code{sequential} does sequential (leave-future-out)
prediction.}

\item{newdata}{Data frame containing the same columns supplied in the
original model.}

\item{xnew}{New predictor matrix or vector. Not required if \code{newdata} is supplied.}

\item{popnew}{New population vector. Not required if \code{newdata} is supplied.}

\item{timenew}{New time vector. Not required if \code{newdata} is supplied.}

\item{ynew}{New response vector. Optional, unless \code{E} and \code{tau} were supplied in
lieu of \code{x}. Not required if \code{newdata} is supplied.}

\item{returnGPgrad}{Return the gradient (derivative) of the GP model at each time point with
respect to each input. This is only computed for out-of-sample predictions using \code{newdata},
\code{loo}, or \code{lto}. If you want the in-sample gradient, pass the original dataset as
\code{newdata}. Defaults to FALSE.}

\item{exclradius}{For \code{predictmethod="loo"} and \code{predictmethod="lto"}, the number of points
on either side of the focal point to leave out of the training data. Defaults to 0.}
}
\value{
A list (class GP and GPpred) with the following elements:
\item{pars}{Posterior mode of hyperparameters (named vector).}
\item{parst}{Posterior mode of hyperparameters (named vector), transformed to real line (used for optimization).}
\item{grad}{Likelihood gradients of hyperparameters at posterior mode. Can be used to assess convergence.}
\item{covm}{List containing various covariance matrices and the inverse covariance matrix.}
\item{iter}{Number of iterations until convergence was reached. Currently fixed at a max of 200.}
\item{inputs}{Inputs and scaled inputs. Note that if you use \code{E} and \code{tau},
the names of the predictors in the input data frame will be stored under
\code{x_names}, and the names of the lagged predictors corresponding to the
inverse length scales will be stored under \code{x_names2}, provided these names exist.}
\item{scaling}{Scaling information.}
\item{insampresults}{Data frame with in-sample predictions. \code{predfsd} is the standard
deviation of the GP function, \code{predsd} includes process error.}
\item{insampfitstats}{Fit statistics for in-sample predictions. Includes R2, rmse, ln_post
(log posterior likelihood evaluated at the mode), lnL_LOO (generalized cross-validation approximate
leave-one-out negative log likelihood), and df (estimated degrees of freedom, equal to
the trace of the smoother matrix). lnL_LOO does not include the prior, so is not directly
comparable to ln_post. For diagnostics, also includes likelihood components ln_prior, SS, logdet.}
\item{insampfitstatspop}{If >1 population, fit statistics (R2 and rmse) for in-sample predictions by population.}
\item{outsampresults}{Data frame with out-of-sample predictions (if requested). \code{predfsd} is the standard
deviation of the GP function, \code{predsd} includes process error.}
\item{outsampfitstats}{Fit statistics for out-of-sample predictions (if requested).
Only computed if using \code{"loo"} or \code{"sequential"}, if \code{y} is found in \code{newdata},
or if \code{ynew} supplied (i.e. if the observed values are known).}
\item{outsampfitstatspop}{If >1 population, fit statistics for out-of-sample predictions (if requested) by population.}
\item{GPgrad}{If \code{returnGPgrad=T}, a data frame with the partial derivatives of the
function with respect to each input.}
\item{call}{Function call. Allows use of \code{\link{update}}.}
}
\description{
Fits a GP model with separable length scales and automatic relevance determination (ARD).
Also fits a hierarchical version of the GP model if distinct populations are indicated
using \code{pop}.
There are several ways to specify the training data:
\itemize{
\item A. supply data frame \code{data}, plus column names or indices for \code{y}, \code{x}, \code{pop}, and \code{time}.
\item B. supply vectors for \code{y}, \code{pop}, and \code{time}, and a matrix or vector for \code{x}.
}
For each of the above 2 options, there are 3 options for specifying the predictors.
\enumerate{
\item supplying \code{y} and \code{x} (omitting \code{E} and \code{tau}) will use the columns of \code{x} as predictors. This allows
for the most customization.
\item supplying \code{y}, \code{E}, and \code{tau} (omitting \code{x}) will use \code{E} lags of \code{y} (with spacing \code{tau}) as predictors.
This is equivalent to option 3 with \code{x}=\code{y}.
\item supplying \code{y}, \code{x}, \code{E}, and \code{tau} will use \code{E} lags of \emph{each column} of \code{x} (with spacing \code{tau}) as predictors.
Do not use this option if \code{x} already contains lags, in that case use option 1.
}
Arguments \code{pop} and \code{time} are optional in all of the above cases.
If \code{pop} and \code{time} are supplied, they should NOT contain missing values.
This function will also, optionally, produce predictions.
See Details for more information.
}
\details{
The data are assumed to be in time order. This is particularly important if \code{E} and \code{tau}
are supplied or \code{predictmethod="sequential"} is used.

\strong{Missing values:}

Missing values for \code{y} and \code{x} are allowed. Any rows containing
missing values will be excluded from the model fitting procedure (reinserted
as NAs in the output). If \code{pop} and \code{time} are supplied, they
should NOT contain missing values.

\strong{Hyperparameters:}

The model uses a squared exponential covariance function. See the (not yet written)
vignette for mathematical details.

There is one inverse length scale \code{phi} estimated for each predictor
variable (input dimension). Values near 0 indicate that the predictor has
little influence on the response variable, and it was likely dropped by ARD.
Higher values of \code{phi} indicate greater nonlinearity.

The estimated variance terms are \code{ve} (process variance) and
\code{sigma2} (pointwise prior variance).

Hyperparameter \code{rho} is the dynamic correlation in the hierarchical GP
model, indicating similarity of dynamics across populations. If there is only
1 population (e.g. if \code{pop} is not supplied), \code{rho} is irrelevant
(not used by the model) and will revert to the mode of its prior (~0.5). The
value of \code{rho} can be fixed to any value from 0.0001 to 0.9999 using
\code{rhofixed}, otherwise it is estimated from the data. Alternatively, a matrix
of fixed pairwise rho values can be supplied using \code{rhomatrix}. In this case,
the single rho parameter will also not be used and will revert to the mode of the
prior (~0.5). A pairwise rho matrix can be generated using \code{\link{getrhomatrix}},
or you can create a custom one (e.g. based on geographical distance).

\strong{Scaling:}

The model priors assume the response variable and all predictor variables
have approximately zero mean and unit variance, therefore it is important to
scale the data properly when fitting these models. For convenience, this
function will scale the input data automatically by default, and
backtransform output to the original scale. Automatic scaling can be
\code{"global"} (default), or \code{"local"}. The latter will scale variables
within (as opposed to across) populations, so is obviously only applicable if
there is more than 1 population. You can also scale the data yourself
beforehand in whatever manner you wish and set \code{scaling="none"}. In this
case, you will obviously have to do the backtransformation yourself.

\strong{Predictions:}

There are several options for producing predictions:
\enumerate{
\item Use \code{predictmethod="loo"} for leave-one-out prediction using the training data.
\item Use \code{predictmethod="lto"} for leave-timepoint-out prediction using the training data. This will leave
out values with the same time index across multiple populations, rather than each individual datapoint.
If there is only one population, \code{"lto"} will be equivalent to \code{"loo"}.
\item Use \code{predictmethod="sequential"} for sequential (leave-future-out) prediction using the training data.
\item If data frame \code{data} was supplied, supply data frame \code{newdata} containing same column names.
Column for \code{y} is optional, unless \code{E} and \code{tau} were supplied in lieu of \code{x}.
\item If vectors/matrices were supplied for \code{y}, \code{x}, etc, equivalent vector/matrices \code{xnew},
\code{popnew} (if \code{pop} was supplied), and \code{timenew} (optional).
\code{ynew} is optional, unless \code{E} and \code{tau} were supplied in lieu of \code{x}.
}

After fitting, predictions can also be made using \code{\link{predict.GP}}.
Predictions can plotted using \code{\link{plot.GPpred}}.
Get conditional responses using \code{\link{getconditionals}}.

It should be noted that \code{"loo"} is not a "true" leave-one-out, for although each prediction is
made by removing one of the training points, the hyperparameters are fit using all of the training data.
The same goes for \code{"sequential"}.

For the out-of-sample prediction methods \code{"loo"}, \code{"lto"}, and
\code{newdata}, the partial derivatives of the fitted GP function at each
time point with respect to each input can be obtained by setting
\code{returnGPgrad = T}. If you want the in-sample gradient, pass the
original (training) data as back in as \code{newdata}. It automatic scaling
was used, the gradient will also be backtransformed to the original units.
}
\examples{
yrand=rnorm(20)
testgp=fitGP(y=yrand,E=2,tau=1,predictmethod = "loo")
}
\references{
Munch, S. B., Poynor, V., and Arriaza, J. L. 2017. Circumventing structural uncertainty:
a Bayesian perspective on nonlinear forecasting for ecology. Ecological Complexity, 32: 134.
}
\seealso{
\code{\link{predict.GP}}, \code{\link{plot.GPpred}}, \code{\link{getconditionals}}, \code{\link{getrhomatrix}}
}
\keyword{functions}
