% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPfunctions.R
\name{makelags}
\alias{makelags}
\title{Generate delay vectors}
\usage{
makelags(
  data = NULL,
  y,
  pop = NULL,
  E,
  tau,
  yname = NULL,
  forecast = FALSE,
  vtimestep = FALSE,
  x = NULL,
  time = NULL,
  augment = FALSE,
  Tdiff_max = NULL,
  Tdiff_fore = NULL,
  nreps = NULL,
  append = FALSE
)
}
\arguments{
\item{data}{A data frame, or matrix with named columns.}

\item{y}{A vector containing a time series (usually the response variable).
Alternatively, a matrix or data frame where each column is a time series
(usually the response variable and covariates). In the case of multiple
time series, lags will be generated for each variable. If \code{data} is
supplied, the column names or indices.}

\item{pop}{A vector of populations (optional, if not supplied defaults to 1
population). If \code{data} is supplied, the column name or index.}

\item{E}{Embedding dimension. Required.}

\item{tau}{Time delay. Required.}

\item{yname}{Optional, name of the variable if \code{y} is input as a vector, or variables
if \code{y} is a matrix with unnamed columns. Otherwise not used.
Defaults to "var" if NULL.}

\item{forecast}{Produce a forecast matrix instead of the standard training/test matrix.}

\item{vtimestep}{Use variable timestep method.}

\item{x}{If using \code{vtimestep=TRUE}, put the response variable under use \code{y}, and
covariates under \code{x} (see details). If \code{vtimestep=FALSE}, will be appended
to \code{y}.}

\item{time}{Used to generate forecast time if \code{forecast=TRUE} and to calculate
Tdiff if \code{vtimestep=TRUE}. If not supplied and \code{vtimestep=TRUE}, a numeric index is used.}

\item{augment}{If \code{vtimestep=TRUE}, produce augmentation lag matrix.}

\item{Tdiff_max}{If \code{vtimestep=TRUE}, the maximum Tdiff value to allow in production
of the lag matrix or augmentation lag matrix.}

\item{Tdiff_fore}{If \code{vtimestep=TRUE} and \code{forecast=TRUE}, vector of time units to
forecast beyond the last timestep. Defaults to the minimum time difference between consecutive
timepoints, multipled by tau.}

\item{nreps}{If \code{augment=TRUE}, the max number of delay vectors for each
Tdiff value.}

\item{append}{Return \code{data} with the lags appended (defaults to FALSE). Only
relevant if \code{forecast=FALSE} and \code{augment=FALSE}.}
}
\value{
A matrix with named columns, the appended number indicating the time lag. If
\code{y} has named columns, named columns in the lag matrix will match. If
generating forecasts, the output matrix will include a column for
population if there is more than one, and will include a time column if
\code{time} is supplied. If using the variable timestep method, the output matrix
will include lags of Tdiff (time difference).
}
\description{
Create a lag matrix for a time series (or suite of time series) using given E
and tau values. Like in fitGP, either a data frame and column names or vectors/matrices
can be provided. Use \code{forecast=TRUE} to generate a forecast matrix. Use
\code{vtimestep=TRUE} to use the variable timestep method.
}
\details{
When using the standard (fixed timestep) method, the response variable and
covariates can all be input under \code{y}, and lags will be generated for
all variables.

When using the variable timestep method, it is necessary to differentiate
between the response variable and covariates, as they are handled
differently. The response variable should go under \code{y} and the
covariates should go under \code{x}. The output matrix will include lags of
Tdiff (time difference).

An augmentation matrix for use with the variable timestep method can be
generated by setting \code{vtimestep=TRUE} and \code{augment=TRUE}. Under
default behavior, the augmentation matrix will include only the Tdiff
combinations observed in the original vtimestep matrix, up to \code{nreps}.
If you supply a vector \code{Tdiff_fore}, then the augmentation matrix will
include or all possible combinations of the Tdiff values supplied in
\code{Tdiff_fore}, even if they weren't in the original vtimestep matrix,
up to \code{nreps}.

If generating forecasts, the output matrix will include a column for
population if there is more than one, and will include a time column if
\code{time} is supplied. The time increment is based on the minimum
difference between timepoints. If generating forecasts using the variable
timestep method, a vector of time units to forecast beyond the last timestep
can be provided under \code{Tdiff_fore}.
}
\examples{
yrand <- rnorm(20)
site <- rep(c("a","b"),each=10)
dfrand <- data.frame(firstvar=rnorm(20),secondvar=rnorm(20))
makelags(y=yrand,E=3,tau=1)
makelags(y=dfrand,E=2,tau=2)
makelags(y=dfrand,pop=site,E=2,tau=1)
makelags(y=yrand,pop=site,E=2,tau=3,forecast = TRUE, yname="SomeName",time=c(1:10,1:10))
dfrand2 <- cbind.data.frame(Time=c(1:10,1:10),Site=site,dfrand)
makelags(data=dfrand2, y=c("firstvar","secondvar"),pop="Site",E=2,tau=3)
makelags(data=dfrand2, y=c("firstvar","secondvar"),pop="Site",E=2,tau=3,
forecast = TRUE,time="Time")

}
