% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPfunctions.R
\name{makelags}
\alias{makelags}
\title{Generate delay vectors}
\usage{
makelags(
  data = NULL,
  yd,
  pop = NULL,
  E,
  tau,
  yname = NULL,
  forecast = FALSE,
  vtimestep = FALSE,
  xd = NULL,
  time = NULL,
  augment = FALSE,
  Tdiffmax = NULL,
  nreps = NULL
)
}
\arguments{
\item{data}{A data frame, or matrix with named columns.}

\item{yd}{A vector containing a time series. Alternatively, a matrix or data
frame where each column is a time series. In the case of multiple time
series, lags will be generated for each variable. If \code{data} is
supplied, the column names or indices.}

\item{pop}{A vector of populations (optional, if not supplied defaults to 1
population). If \code{data} is supplied, the column name or index.}

\item{E}{Embedding dimension.}

\item{tau}{Time delay.}

\item{yname}{Optional, name of the variable if yd is input as a vector, or variables
if yd is a matrix with unnamed columns. Otherwise not used.
Defaults to "var" if NULL.}

\item{forecast}{Produce a forecast matrix instead of the standard training/test matrix.}

\item{vtimestep}{Use variable timestep method.}

\item{xd}{If using \code{vtimestep=TRUE}, put the response variable under use \code{yd}, and
covariates under \code{xd} (see details)}

\item{time}{Used to generate forecast time if \code{forecast=TRUE} and to calculate
Tdiff if \code{vtimestep=TRUE}. If not supplied and \code{vtimestep=TRUE}, a numeric index is used.}

\item{augment}{If \code{vtimestep=TRUE}, produce augmentation lag matrix.}

\item{Tdiffmax}{If \code{vtimestep=TRUE}, the maximum Tdiff value to allow in production
of the lag matrix or aumentation lag matrix.}

\item{nreps}{If \code{augment=TRUE}, the max number of delay vectors for each
Tdiff value.}
}
\value{
A matrix with named columns, the appended number indicating tau. If
\code{yd} has named columns, named columns in the lag matrix will match. If
generating forecasts, the output matrix will include a column for
population if there is more than one, and will include a time column if
\code{time} is supplied. If using the variable timestep method, the output matrix
will include lags of Tdiff (time difference).
}
\description{
Create a lag matrix for a time series (or suite of time series) using given E
and tau values. Like in fitGP, either a data frame and column names or vectors/matrices
can be provided. Use \code{forecast=TRUE} to generate a forecast matrix. Use
\code{vtimestep=TRUE} to use the variable timestep method.
}
\details{
When using the standard (fixed timestep) method, the response variable and
covariates can all be input under \code{yd}, and lags will be generated for
all variables.

When using the variable timestep method, and it is necessary to differentiate
between the response variable and covariates, as they are handled
differently. The response variable should go under \code{yd} and the
covariates should go under \code{xd}. The output matrix will include lags of
Tdiff (time difference).

An augmentation matrix for use with the variable timestep method can be
generated by setting \code{vtimestep=TRUE} and \code{augment=TRUE}.

If generating forecasts, the output matrix will include a column for population if there
is more than one, and will include a time column if \code{time} is supplied. The time
increment is based on the difference between the last and second to last timepoints.
If generating forecast using the variable timestep method...
}
\examples{
yrand <- rnorm(20)
site <- rep(c("a","b"),each=10)
dfrand <- data.frame(firstvar=rnorm(20),secondvar=rnorm(20))
makelags(yd=yrand,E=3,tau=1)
makelags(yd=dfrand,E=2,tau=2)
makelags(yd=dfrand,pop=site,E=2,tau=1)
makelags(yd=yrand,pop=site,E=2,tau=3,forecast = T, yname="SomeName",time=c(1:10,1:10))
dfrand2 <- cbind.data.frame(Time=c(1:10,1:10),Site=site,dfrand)
makelags(data=dfrand2, yd=c("firstvar","secondvar"),pop="Site",E=2,tau=3)
makelags(data=dfrand2, yd=c("firstvar","secondvar"),pop="Site",E=2,tau=3,forecast = T,time="Time")

}
