% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotsConditionals.R
\name{plot.NStest}
\alias{plot.NStest}
\title{Plot results from an S-map nonstationarity test}
\usage{
\method{plot}{NStest}(x, ...)
}
\arguments{
\item{x}{Output from \code{Smap_NStest}.}

\item{...}{Other graphical parameters for the time series plot, like 'main'.}
}
\description{
Plots some of the results from the S-map nonstationarity test. The first plot
is the time series. The others are the logL, delta value, weights, degrees of freedom,
and leave-one-out R2 as functions of E.
}
\keyword{functions}
