% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhomatrix.R
\name{posdef}
\alias{posdef}
\title{Make a matrix postive definite}
\usage{
posdef(mat)
}
\arguments{
\item{mat}{A symmetrical square matrix.}
}
\value{
A version of \code{mat} that is positive definite.
}
\description{
If the output matrix from \code{\link{getrhomatrix}} is not positive
definite, this will convert it to a matrix that is.
}
\seealso{
\code{\link{getrhomatrix}}
}
\keyword{functions}
